EPIC 1:
The code has been fixed to better fit the mock-up.
The genre can now have a parent on the back-end, and on the front-end, it now appears in a tree structure.
The searchFilter controller has been fixed to allow searching by name and genre.
Now there is only one "Change Game" button for making changes.
On the game add page, images can be added, and the default image is now taken from the front-end img folder instead of the back-end (other non-default images are still saved in the wwwroot/img folder on the back-end).
Additionally, there are some other minor front-end fixes.